requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
fastmcp>=2.4.0
google-api-python-client>=2.86.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
python-dotenv>=1.0.0
